+++ test/branch-deps.js
@@ -15,12 +15,12 @@ goog.addDependency("../../../../../src/3_branch.js", ['Branch'], ['utils', 'reso
 goog.addDependency("../../../../../src/4_initialization.js", ['branch_instance'], ['Branch', 'config']);
 goog.addDependency("../../../../../src/extern.js", [], []);
 goog.addDependency("../../../../../src/onpage.js", [], []);
-goog.addDependency("../../../../../src/titanium/BranchSDK/example/app.js", [], []);
 goog.addDependency("../../../../../test/0_queue.js", [], ['Queue']);
 goog.addDependency("../../../../../test/0_storage.js", [], ['storage']);
 goog.addDependency("../../../../../test/0_utils.js", [], ['utils']);
 goog.addDependency("../../../../../test/1_api.js", [], ['utils', 'Server', 'resources', 'storage']);
 goog.addDependency("../../../../../test/3_branch.js", [], ['utils', 'Branch', 'resources', 'config', 'storage', 'web_session', 'goog.json']);
 goog.addDependency("../../../../../test/4_integration.js", [], ['config', 'goog.json']);
+goog.addDependency("../../../../../test/saucelabs.js", [], []);
 goog.addDependency("../../../../../test/test-utils.js", [], []);
 goog.addDependency("../../../../../test/web-config.js", ['config'], []);